Environment Work
By using environments, we can collect a number of objects together and save these environments to files.
With the hold environment, all of the work done by the Data.r script can be removed, but first it puts dataALL into the Global Environment.
Later the saveRDS function can be used to save the DATA environment for loading later. (Commented out as not all of this design has been implemented.)