More accurate use of cut
Changing cut's arguments so that it will be closed on the left and open on the right. (Include 0 in the test period and exclude 3600, making that the 0 for Cooldown.) It had included both extremes in the Test period but now just one.
While this change is more accurate, data does not need to be recompiled. The difference is a single second and thanks to thermal inertia and the time to actually close programs, this is likely academic