Moved calls for Output Generation
Pulled the commands that actually generate the output files out of the Output.r script and put them in Input.r. The reason for this is to make selectively generating these outputs easier. For example, if a change has been made to one graph, this will make it easier to generate that one, rather than them all over again.
This mirrors a change I have made with the OCC-CPU_Thermals scripts.